+++ src/Exception/NoMatchingChangelogDiscoveredException.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * @see       https://github.com/phly/keep-a-changelog-tagger for the canonical source repository
+ * @copyright Copyright (c) 2018 Matthew Weier O'Phinney
+ * @license   https://github.com/phly/keep-a-changelog-tagger/blob/master/LICENSE.md New BSD License
+ */
+
+declare(strict_types=1);
+
+namespace Phly\KeepAChangelog\Exception;
+
+use RuntimeException;
+
+class NoMatchingChangelogDiscoveredException extends RuntimeException
+{
+    public static function for(string $changelogFile) : self
+    {
+        return new self(sprintf(
+            'Unable to find un-dated changelog entry in %s',
+            $changelogFile
+        ));
+    }
+}